refactor(capabilities): split the http client and server caps into identity and runtime modules
Closes #2323.
